sms-service/requirements.txt
flask==2.0.1
requests==2.26.0
numpy==1.21.2
scikit-learn==0.24.2
tensorflow==2.6.0
python-dotenv==0.19.0
pytest==6.2.5
pandas==1.3.3
joblib==1.0.1 